+++ components/catalogs/back/tests/Integration/Infrastructure/Controller/Public/GetProductUuidsActionTest.php
@@ -61,7 +61,7 @@ class GetProductUuidsActionTest extends IntegrationTestCase
     public function testItReturnsAnErrorMessagePayloadWhenTheCatalogIsDisabled(): void
     {
         $this->client = $this->getAuthenticatedPublicApiClient(['read_catalogs', 'read_products']);
-        $this->createCatalog('db1079b6-f397-4a6a-bae4-8658e64ad47c', 'Store US', 'shopifi');
+        $this->createCatalog('db1079b6-f397-4a6a-bae4-8658e64ad47c', 'Store US', 'shopifi', false);
         $this->createProduct('blue');
 
         $this->client->request(