+++ components/catalogs/back/tests/Integration/Infrastructure/Persistence/Catalog/FindOneCatalogByIdQueryTest.php
@@ -32,7 +32,7 @@ class FindOneCatalogByIdQueryTest extends IntegrationTestCase
         $this->createUser('test');
         $id = 'db1079b6-f397-4a6a-bae4-8658e64ad47c';
 
-        $this->createCatalog($id, 'Store US', 'test');
+        $this->createCatalog($id, 'Store US', 'test', isEnabled: false);
 
         $result = $this->query->execute($id);