+++ components/catalogs/front/src/components/ProductMapping/components/SourcePanel.tsx
@@ -5,16 +5,23 @@ import {SelectAttributeDropdown} from './SelectAttributeDropdown';
 import {useTranslate} from '@akeneo-pim-community/shared';
 import {Source} from '../models/Source';
 import {Attribute} from '../../../models/Attribute';
+import {SelectChannelDropdown} from './SelectChannelDropdown';
+import {useAttribute} from '../../../hooks/useAttribute';
+import {SourceErrors} from '../models/SourceErrors';
+import {SelectLocaleDropdown} from './SelectLocaleDropdown';
+import {SelectChannelLocaleDropdown} from './SelectChannelLocaleDropdown';
 
 type Props = {
     target: string | null;
     source: Source | null;
     targetLabel: string | null;
     onChange: (value: Source) => void;
+    errors: SourceErrors | null;
 };
 
-export const SourcePanel: FC<Props> = ({target, source, targetLabel, onChange}) => {
+export const SourcePanel: FC<Props> = ({target, source, targetLabel, onChange, errors}) => {
     const translate = useTranslate();
+    const {data: attribute} = useAttribute(source?.source ?? '');
     const handleSourceSelection = useCallback(
         (value: Attribute) => {
             onChange({
@@ -25,6 +32,11 @@ export const SourcePanel: FC<Props> = ({target, source, targetLabel, onChange}) 
         },
         [onChange]
     );
+
+    const shouldDisplayChannel = source !== null && attribute?.scopable;
+    const shouldDisplayLocale = source !== null && attribute?.localizable && !attribute?.scopable;
+    const shouldDisplayChannelLocale =
+        source !== null && source.scope !== null && attribute?.localizable && attribute?.scopable;
 
     return (
         <>
@@ -41,9 +53,25 @@ export const SourcePanel: FC<Props> = ({target, source, targetLabel, onChange}) 
                         </SectionTitle.Title>
                     </SectionTitle>
                     <SelectAttributeDropdown
-                        code={null !== source && null !== source.source ? source.source : ''}
+                        code={source?.source ?? ''}
                         onChange={handleSourceSelection}
+                        error={errors?.source}
                     ></SelectAttributeDropdown>
+                    <SectionTitle>
+                        <Tag tint='purple'>2</Tag>
+                        <SectionTitle.Title level='secondary'>
+                            {translate('akeneo_catalogs.product_mapping.source.parameters.title')}
+                        </SectionTitle.Title>
+                    </SectionTitle>
+                    {shouldDisplayChannel && (
+                        <SelectChannelDropdown source={source} onChange={onChange} error={errors?.scope} />
+                    )}
+                    {shouldDisplayLocale && (
+                        <SelectLocaleDropdown source={source} onChange={onChange} error={errors?.locale} />
+                    )}
+                    {shouldDisplayChannelLocale && (
+                        <SelectChannelLocaleDropdown source={source} onChange={onChange} error={errors?.locale} />
+                    )}
                 </>
             )}
         </>