+++ components/catalogs/front/src/components/ProductValueFilters/models/Locale.d.ts
@@ -1,4 +1,0 @@
-export type Locale = {
-    label: string;
-    code: string;
-};