+++ components/identifier-generator/back/src/Domain/Exception/UnableToSetIdentifierException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Akeneo\Pim\Automation\IdentifierGenerator\Domain\Exception;
-
-class UnableToSetIdentifierException extends \Exception
-{
-}