+++ components/identifier-generator/front/src/feature/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './useGetIdentifierGenerators';
 export * from './useIdentifierAttributes';
 export * from './useUiLocales';
 export * from './useValidateFormWithEnter';
+export * from './useStructureTabs';